# routes/authRoutes.js
const express = require("express")
const bcrypt = require("bcryptjs")
const jwt = require("jsonwebtoken")
const User = require("../models/User.model")
const router = express.Router()
router.post("/register", async (req, res) => {
  try {
    const { username, email, password, role } = req.body
    const hashedPassword = await bcrypt.hash(password, 10)
    const newUser = new User({ username, email, password: hashedPassword, role })
    await newUser.save()
    res.json({ message: "User registered successfully" })
  } catch (err) {
    res.status(500).json({ message: "Error registering user", error: err })
  }
})
router.post("/login", async (req, res) => {
  try {
    const { email, password } = req.body
    const user = await User.findOne({ email })
    if (!user) return res.status(400).json({ message: "User not found" })
    const validPassword = await bcrypt.compare(password, user.password)
    if (!validPassword) return res.status(400).json({ message: "Invalid password" })
    const token = jwt.sign({ _id: user._id }, process.env.JWT_SECRET, { expiresIn: "1h" })
    res.json({ token })
  } catch (err) {
    res.status(500).json({ message: "Error logging in", error: err })
  }
})
module.exports = router;